improving data structure for response and request
